server/requirements.txt
lxml~=4.8.0
beautifulsoup4~=4.10.0
soupsieve~=2.3.1
pip~=22.0.3
Jinja2~=3.0.3
future~=0.18.2
py4j~=0.10.9
pytz~=2021.3
MarkupSafe~=2.0.1
Werkzeug~=2.0.3
click~=8.0.3
itsdangerous~=2.0.1
neo4j~=4.4.1
pansi~=2020.7.3
six~=1.16.0
py2neo~=2021.2.3
interchange~=2021.0.4
packaging~=21.3
urllib3~=1.26.8
monotonic~=1.6
Pygments~=2.11.2
thrift~=0.15.0
certifi~=2021.10.8
graphviz~=0.19.1
idna~=3.3
Chameleon~=3.9.1
pyparsing~=3.0.7
setuptools~=60.9.1
requests~=2.27.1
tabulate~=0.8.9
Flask~=2.0.2
supervisor
PyYAML

